I actually believe parseFloat is correct.
Because the acceleration is constant (when using the keyboard), the
velocity graph should be a straight line.

+++ source.js
@@ -48,8 +48,8 @@ function toggleTimer(){  //Called from HTML
 
 function myTimer() {
 	moving += speed;
-	$('#slider').text(parseInt(document.getElementById('slider').innerHTML)+moving);  //Update the slider's text
-	var sliderPosition = parseInt(document.getElementById('slider').innerHTML);
+	$('#slider').text(parseFloat(document.getElementById('slider').innerHTML)+moving);  //Update the slider's text
+	var sliderPosition = parseFloat(document.getElementById('slider').innerHTML);
 	//Drawing position
 	drawPos(-sliderPosition+250);  //Invert the position so that it displays properly then add 250 because that's the center of the graph
 	timeCounter+=1;
